modularized code some, waiting until saturday to fix final bugs before finish modularization
